% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{rgnorm}
\alias{rgnorm}
\title{Generalized Normal Distribution.}
\usage{
rgnorm(n = 1, mu = 0, alpha = 1, beta = 1)
}
\arguments{
\item{n}{Number of generated samples.}

\item{mu}{mean.}

\item{alpha}{first shape parameter.}

\item{beta}{second shape parameter.}
}
\value{
numeric array of length \eqn{n}.
}
\description{
Random generation for generalized Normal Distribution.
}
\keyword{internal}
