% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{rlaplace}
\alias{rlaplace}
\title{Laplace distribution}
\usage{
rlaplace(n = 1, mu = 0, sd = 1)
}
\arguments{
\item{n}{Number of generated samples.}

\item{mu}{mean.}

\item{sd}{standard deviation.}
}
\value{
numeric array of length \eqn{n}.
}
\description{
Random generation for Laplace distribution.
}
\keyword{internal}
